@@ -6,13 +6,22 @@
     "@testing-library/jest-dom": "^5.12.0",
     "@testing-library/react": "^11.2.6",
     "@testing-library/user-event": "^12.8.3",
+    "@types/classnames": "^2.3.1",
     "@types/jest": "^26.0.23",
     "@types/node": "^12.20.10",
-    "@types/react": "^17.0.4",
-    "@types/react-dom": "^17.0.3",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "@types/react": "^18.0.35",
+    "@types/react-dom": "^18.0.11",
+    "@types/react-router-dom": "^5.3.3",
+    "@types/react-transition-group": "^4.4.5",
+    "classnames": "^2.3.2",
+    "localforage": "^1.10.0",
+    "match-sorter": "^6.3.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.10.0",
     "react-scripts": "4.0.3",
+    "react-transition-group": "^4.4.5",
+    "sort-by": "^1.2.0",
     "typescript": "^4.2.4",
     "typescript-eslint-parser": "^22.0.0",
     "web-vitals": "^1.1.1"
